+++ packages/perseus/src/widgets/categorizer/categorizer.stories.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 
 import {RendererWithDebugUI} from "../../../../../testing/renderer-with-debug-ui";
-import {question1} from "../__testdata__/categorizer.testdata";
+
+import {question1} from "./categorizer.testdata";
 
 export default {
     title: "Perseus/Widgets/Categorizer",
