+++ packages/perseus/src/widgets/categorizer/categorizer.tsx
@@ -5,20 +5,20 @@ import classNames from "classnames";
 import * as React from "react";
 import _ from "underscore";
 
-import {PerseusI18nContext} from "../components/i18n-context";
-import InlineIcon from "../components/inline-icon";
-import {iconCircle, iconCircleThin} from "../icon-paths";
-import * as Changeable from "../mixins/changeable";
-import WidgetJsonifyDeprecated from "../mixins/widget-jsonify-deprecated";
-import {ClassNames as ApiClassNames} from "../perseus-api";
-import Renderer from "../renderer";
-import mediaQueries from "../styles/media-queries";
-import sharedStyles from "../styles/shared";
-import Util from "../util";
+import {PerseusI18nContext} from "../../components/i18n-context";
+import InlineIcon from "../../components/inline-icon";
+import {iconCircle, iconCircleThin} from "../../icon-paths";
+import * as Changeable from "../../mixins/changeable";
+import WidgetJsonifyDeprecated from "../../mixins/widget-jsonify-deprecated";
+import {ClassNames as ApiClassNames} from "../../perseus-api";
+import Renderer from "../../renderer";
+import mediaQueries from "../../styles/media-queries";
+import sharedStyles from "../../styles/shared";
+import Util from "../../util";
 
-import type {PerseusCategorizerWidgetOptions} from "../perseus-types";
-import type {PerseusStrings} from "../strings";
-import type {PerseusScore, WidgetExports, WidgetProps} from "../types";
+import type {PerseusCategorizerWidgetOptions} from "../../perseus-types";
+import type {PerseusStrings} from "../../strings";
+import type {PerseusScore, WidgetExports, WidgetProps} from "../../types";
 
 type UserInput = any;
