+++ packages/perseus/src/widgets/categorizer/index.ts
@@ -1,0 +1,1 @@
+export {default, Categorizer} from "./categorizer";
